requirements.txt
trafilatura>=2.0.0
requests>=2.32.0
lxml>=6.0.0
playwright>=1.57.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
